medvqa (CNN Model)/requirements.txt
torch
torchvision
transformers
datasets
tqdm
numpy
pillow
gradio
pyyaml
